+++ src/app/app.module.ts
@@ -14,7 +14,13 @@ import { environment } from 'src/environments/environment';
 import { AngularFireModule } from '@angular/fire';
 import { AngularFirestoreModule } from '@angular/fire/firestore';
 import { AngularFireAuthModule } from '@angular/fire/auth';
-import { NotFoundComponent } from './components/not-found.component';
+
+// ngrx
+import { StoreModule } from '@ngrx/store';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { appReducers } from './store/app.reducers';
+
+import { NotFoundComponent } from './components';
 
 @NgModule({
   declarations: [
@@ -28,7 +34,12 @@ import { NotFoundComponent } from './components/not-found.component';
     RouterModule.forRoot(appRputes),
     AngularFireModule.initializeApp(environment.firebase),
     AngularFirestoreModule,
-    AngularFireAuthModule
+    AngularFireAuthModule,
+    StoreModule.forRoot(appReducers),
+    StoreDevtoolsModule.instrument({
+      maxAge: 25, // Retains last 25 states
+      logOnly: environment.production, // Restrict extension to log-only mode
+    })
   ],
   bootstrap: [AppComponent]
 })