+++ src/app/components/auth/login.component.ts
@@ -1,17 +1,23 @@
 import { Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { AuthService } from 'src/app/services';
+import { Store } from '@ngrx/store';
+import { AppState } from 'src/app/store/app.reducers';
+import { Observable } from 'rxjs';
+import { UiState } from 'src/app/store/reducers';
 
 @Component({
   selector: 'app-login.auth-form',
   template: /*html*/`
+    <mat-progress-bar mode="query" *ngIf="(loadingState$ | async).isLoadig"></mat-progress-bar>
     <mat-card class="flex-column" [formGroup]="form">
       <form class="flex-column"
         [formGroup]="form"
         (ngSubmit)="onSubmit()">
         <mat-form-field>
+          <mat-label>Email</mat-label>
+          <mat-icon matPrefix class="icon-prefix">email</mat-icon>
           <input matInput
-            placeholder="email"
             type="email"
             formControlName="email">
 
@@ -20,9 +26,16 @@ import { AuthService } from 'src/app/services';
         </mat-form-field>
 
         <mat-form-field>
+          <mat-label>Password</mat-label>
+          <button class="icon-prefix"
+            mat-icon-button
+            matPrefix
+            (click)="togglePassword()"
+            type="button">
+            <mat-icon>{{ passwordVisibility ? 'visibility_off' : 'visibility' }} </mat-icon>
+          </button>
           <input matInput
-            placeholder="password"
-            type="password"
+            [type]="passwordVisibility ? 'text' : 'password'"
             formControlName="password"
             autocomplete="off">
 
@@ -36,7 +49,8 @@ import { AuthService } from 'src/app/services';
         <button class="auth-form__submit"
           mat-raised-button
           type="submit"
-          color="primary">
+          color="primary"
+          [disabled]="form.invalid || (loadingState$ | async).isLoadig">
           Login
         </button>
       </form>
@@ -52,14 +66,18 @@ import { AuthService } from 'src/app/services';
 export class LoginComponent implements OnInit {
   public form: FormGroup;
   public formError = null;
+  public loadingState$: Observable<UiState>;
+  public passwordVisibility = false;
 
-  constructor(private auth: AuthService) {}
+  constructor(private auth: AuthService, private store: Store<AppState>) {}
 
   ngOnInit(): void {
     this.form = new FormGroup({
-      email: new FormControl('jorge.barron@amplemind.com', [Validators.email, Validators.required]),
-      password: new FormControl('123123123', Validators.required)
+      email: new FormControl('', [Validators.email, Validators.required]),
+      password: new FormControl('', Validators.required)
     });
+
+    this.loadingState$ = this.store.select('ui');
   }
 
   public onSubmit(): void {
@@ -68,5 +86,9 @@ export class LoginComponent implements OnInit {
 
   public hasError(controlName: string, typeError: string): boolean {
     return this.form.get(controlName).hasError(typeError);
+  }
+
+  public togglePassword() {
+    this.passwordVisibility = !this.passwordVisibility;
   }
 }