+++ src/app/components/auth/register.component.ts
@@ -1,18 +1,24 @@
 import { Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { AuthService } from 'src/app/services';
+import { Observable } from 'rxjs';
+import { UiState } from 'src/app/store/reducers';
+import { Store } from '@ngrx/store';
+import { AppState } from 'src/app/store/app.reducers';
 
 @Component({
   selector: 'app-register.auth-form',
   template: /*html*/`
+    <mat-progress-bar mode="query" *ngIf="(loadingState$ | async).isLoadig"></mat-progress-bar>
     <mat-card>
       <form class="flex-column"
         [formGroup]="form"
         (ngSubmit)="onSubmit()">
         <mat-form-field>
+          <mat-label>Name</mat-label>
+          <mat-icon matPrefix class="icon-prefix">person</mat-icon>
           <input matInput
             autocomplete="off"
-            placeholder="Name"
             type="text"
             formControlName="name">
 
@@ -20,8 +26,9 @@ import { AuthService } from 'src/app/services';
         </mat-form-field>
 
         <mat-form-field>
+          <mat-label>Email</mat-label>
+          <mat-icon matPrefix class="icon-prefix">email</mat-icon>
           <input matInput
-            placeholder="Email"
             type="email"
             formControlName="email">
 
@@ -30,10 +37,17 @@ import { AuthService } from 'src/app/services';
         </mat-form-field>
 
         <mat-form-field>
+          <mat-label>Password</mat-label>
+          <button class="icon-prefix"
+            mat-icon-button
+            matPrefix
+            (click)="togglePassword()"
+            type="button">
+            <mat-icon>{{ passwordVisibility ? 'visibility_off' : 'visibility' }} </mat-icon>
+          </button>
           <input matInput
             autocomplete="off"
-            placeholder="Password"
-            type="password"
+            [type]="passwordVisibility ? 'text' : 'password'"
             formControlName="password">
 
           <mat-error *ngIf="hasError('password', 'required')">The field is required</mat-error>
@@ -47,8 +61,8 @@ import { AuthService } from 'src/app/services';
           mat-raised-button
           type="submit"
           color="primary"
-          [disabled]="form.invalid">
-          Register
+          [disabled]="form.invalid || (loadingState$ | async).isLoadig">
+          {{ (loadingState$ | async).isLoadig ? 'Creating user...' : 'Register' }}
         </button>
       </form>
     </mat-card>
@@ -63,8 +77,10 @@ import { AuthService } from 'src/app/services';
 export class RegisterComponent implements OnInit {
   public form: FormGroup;
   public formError = null;
+  public loadingState$: Observable<UiState>;
+  public passwordVisibility = false;
 
-  constructor(private auth: AuthService) {}
+  constructor(private auth: AuthService, private store: Store<AppState>) {}
 
   ngOnInit(): void {
     this.form = new FormGroup({
@@ -72,6 +88,8 @@ export class RegisterComponent implements OnInit {
       email: new FormControl('', [Validators.email, Validators.required]),
       password: new FormControl('', Validators.required)
     });
+
+    this.loadingState$ = this.store.select('ui');
   }
 
   public onSubmit() {
@@ -84,5 +102,9 @@ export class RegisterComponent implements OnInit {
 
   public hasError(controlName: string, typeError: string): boolean {
     return this.form.get(controlName).hasError(typeError);
+  }
+
+  public togglePassword() {
+    this.passwordVisibility = !this.passwordVisibility;
   }
 }