+++ src/app/modules/auth.module.ts
@@ -1,10 +1,9 @@
 import { NgModule } from '@angular/core';
 import { LoginComponent, RegisterComponent } from '../components';
 import { RouterModule } from '@angular/router';
-import { authRoutes } from '../routes';
 import { SharedModule } from '.';
 import { AuthComponent } from '../components/auth/auth.component';
-import { ReactiveFormsModule } from '@angular/forms';
+import { authRoutes } from '../routes';
 
 @NgModule({
   declarations: [
@@ -15,7 +14,6 @@ import { ReactiveFormsModule } from '@angular/forms';
   imports: [
     SharedModule,
     RouterModule.forChild(authRoutes),
-    ReactiveFormsModule
   ]
 })
 export class AuthModule {}