+++ src/app/store/reducers/index.ts
@@ -1,0 +1,3 @@
+export * from './ui.reducers';
+export * from './auth.reducers';
+export * from './income-expense.reducers';